# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  dto.MergeResponse:
    properties:
      aibom_component_count:
        type: integer
      dataset_components:
        items:
          type: string
        type: array
      duplicates_removed:
        type: integer
      merged_bom:
        items:
          type: integer
        type: array
      metadata_component:
        type: string
      model_components:
        items:
          type: string
        type: array
      sbom_component_count:
        type: integer
      sbom_components:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: IDLab Discover
    url: https://github.com/idlab-discover/AIBoMGen-cli
  description: HTTP API for generating, assessing and merging CycloneDX AI/ML Bills
    of Materials (AIBOMs) using the AIBoMGen-cli library.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AIBoMGen REST API
  version: "1.0"
paths:
...
  /merge/aiboms-with-sbom:
    post:
      consumes:
      - multipart/form-data
      description: Merges one or more AIBOMs into a base SBOM. This is the recommended
        workflow when integrating AI/ML model metadata into an existing software BOM.
      parameters:
      - description: Base SBOM file
        in: formData
        name: sbom
        required: true
        type: file
      - collectionFormat: csv
        description: One or more AIBOM files (select multiple in the file picker)
        in: formData
        items:
          type: file
        name: aibom[]
        required: true
        type: array
      - description: Label for the SBOM in merged metadata
        in: formData
        name: sbom_name
        type: string
      - description: Remove duplicate components (default false)
        in: formData
        name: deduplicate_components
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MergeResponse'
        "400":
          description: Missing required file(s)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: BOM parse error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Merge error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Merge AIBOMs into an SBOM
      tags:
      - merge
...
schemes:
- http
- https
swagger: "2.0"